+++ src/ec4py/analysis_levich.py
@@ -47,11 +47,12 @@ def Levich(rot, y_data, y_axis_unit:str="A", y_axis_title:str="i", STYLE_DL: str
         ##print("AAA",B_factor_pos, "BBB", B_factor_pos.quantity)
         
         #Levich Plot
-        p = plot_options(kwargs)
+        p = plot_options(**kwargs)
         p.no_smooth()
        
         p.set_title("Levich",1)
-        p.set_x_txt("$\omega$^0.5", f"{rot_unit}^0.5")
+        s=r"\omega"
+        p.set_x_txt(f"${s}$^0.5", f"{rot_unit}^0.5")
         p.set_y_txt(y_axis_title, y_axis_unit)
 
         p.options["style"]=STYLE_DL[0]+"o"