+++ src/ec4py/analysis_tafel.py
@@ -32,7 +32,7 @@ def Tafel(x_data, y_data, y_axis_unit, y_axis_title, plot_color, lineName="", x_
 
    
  
-    p = plot_options(kwargs)
+    p = plot_options(**kwargs)
     p.no_smooth()
     
     p.set_title("Tafel")