+++ src/ec4py/step_data.py
@@ -12,11 +12,15 @@ import copy
 
 from .ec_data import EC_Data,index_at_time
 from .ec_data_util import EC_Channels
+from .util_data import get_IR
+
 
 from .ec_setup import EC_Setup
 from .util_graph import plot_options,quantity_plot_fix, make_plot_2x,make_plot_1x,make_plot_2x_1,saveFig
 from .util import extract_value_unit     
 from .util import Quantity_Value_Unit as QV
+ 
+
 from .analysis_tafel import Tafel
 
 
@@ -47,6 +51,7 @@ class Step_Data(EC_Setup):
         self.step_E =[]
         self.step_Type =[]
         self.E_shifted_by = None
+        self.IR_COMPENSATED = False
         if not args:
             return
         else:
@@ -76,7 +81,7 @@ class Step_Data(EC_Setup):
         Returns:
             dict[]: list of key words
         """
-        return self.setup_data._setup["Repetitions"]
+        return self.setup_data._setup.get("Repetitions","1")
     
     @property 
     def nr_of_steps(self):
@@ -85,7 +90,7 @@ class Step_Data(EC_Setup):
         Returns:
             dict[]: list of key words
         """
-        return int(self.setup_data._setup["Repetitions"])*len(self.step_Time)
+        return int(self.repetitions)*len(self.step_Time)
     ##########################       
     def conv(self, ec_data: EC_Data, *args, ** kwargs):
         """Converts EC_Data to a CV
@@ -102,9 +107,12 @@ class Step_Data(EC_Setup):
         options = {
             'x_smooth' : 0,
             'y_smooth' : 0,
-            'IR': 0
+            'IRCORR': None,
+            'E' : "E",
+            'i' : 'i'
         }
         options.update(kwargs)
+      
         sel_channels = EC_Channels(*args,**kwargs)
         try:
             data_E,q_E,u_E,_ = ec_data.get_channel(sel_channels.Voltage)
@@ -113,19 +121,32 @@ class Step_Data(EC_Setup):
             data_P,q,u,_ = ec_data.get_channel(sel_channels.Phase)
             self.setup_data = ec_data.setup_data
             #self.convert(ec_data.Time,ec_data.E,ec_data.i,**kwargs)
+
             self.Time = ec_data.Time
             self.i = data_i
             self.E = data_E
             self.Z = data_Z
             self.P = data_P
             
+            dt = (np.max(self.Time)-np.min(self.Time)) / (len(self.Time)-1)
+            #print("conv_dt", dt)
+            self.step_Time = List_Str2float(self.setup.get("Step.Time",str(f"{np.max(self.Time)+dt};")))
+            self.step_E =List_Str2float(self.setup.get("Step.E",f"{np.mean(data_E)};"))
+            self.step_Type = List_Str2Str(self.setup.get("Step.Type","H"))
+            self.E_label = q_E + " vs " + self.RE
+
+            comp = options.get("IRCORR",None)
+            if comp is not None:
+                ir_comp, data_IR = get_IR(ec_data,sel_channels,comp)
+                if ir_comp:
+                    self.E = data_E - data_IR
+                    self.E_label = q_E + " vs " + self.RE + " - IR"
+                    self.IR_COMPENSATED = True
+               
             #self.step_Time = self.setup["Step.Time"].split(";",-1)
             #self.step_E = self.setup["Step.E"].split(";",-1)
             #self.step_Type = self.setup_data._setup["Step.Type"].split(";",-1)
-            self.step_Time = List_Str2float(self.setup["Step.Time"])
-            self.step_E =List_Str2float(self.setup["Step.E"])
-            self.step_Type = List_Str2Str(self.setup["Step.Type"])
-            self.E_label = q_E + " vs " + self.RE
+            
         except ValueError as e:
             print("no_data",e)
         #self.setup = data.setup
@@ -155,19 +176,20 @@ class Step_Data(EC_Setup):
         range.update(kwargs)
         #print(kwargs)
         #print(range)
-        options = plot_options(kwargs)
+        options = plot_options(**kwargs)
+       
         data = copy.deepcopy(self)
-        
-        # print("plotARGS",args)
+        options.legend = data.legend(x_channel,y_channel,*args, **kwargs)
+        #print("plotARGS",args)
         data.norm(args)
         data.set_active_RE(args)
         # print("QQQQ",data.E_label)
         index_min = 0
         if range["limit_min"] >0:
             index_min = data.index_at_time(range["limit_min"])
-        index_max = len(data.Time)-1
+        index_max = len(data.Time) #-1
         if range["limit_max"] >0:
-            index_max = data.index_at_time(range["limit_max"])
+            index_max = np.min([data.index_at_time(range["limit_max"]),len(data.Time)-1])
         
         #max_index = len(self.Time)-1
         #print("index", index_min,index_max)
@@ -269,7 +291,7 @@ class Step_Data(EC_Setup):
         for i in range(0, total_nr_steps):
             num =float(self.step_Time[i])
             s.append(s[i]+num)
-        #print(s)
+        #print("get_step_step_times",s)
         idx = step_index % total_nr_steps
        # print("total", total_nr_steps)
        # print("idx", idx)
@@ -280,19 +302,21 @@ class Step_Data(EC_Setup):
         startT = s[idx]
         endT = s[idx+step_range]
          
-     #   print("startT",startT)
-      #  print("endT",endT)
+        #print("startT",startT)
+        #print("endT",endT)
         
         
         
         start_index = self.index_at_time(startT+extra)-1
         if start_index< 0 : 
             start_index = 0
-        end_index =   self.index_at_time(endT+extra)-1
+        end_index =   self.index_at_time(endT+extra)  #+1
+        #if this is the last step, add 1 to the end index
+        if endT>= np.max(self.Time): 
+            end_index = end_index + 1
         #shift dateTime
         singleStep.setup_data.dateTime = self.setup_data.dateTime + np.timedelta64(int(self.Time[start_index]*1000000),'us')
-      #  print("startT",start_index)
-     #   print("endT",end_index)
+        # print("startIndex",start_index,"endIndex",end_index)
         aSize=end_index-start_index #+1
         singleStep.E = np.empty(aSize) 
         singleStep.i = np.empty(aSize) 
@@ -323,7 +347,7 @@ class Step_Data(EC_Setup):
         self.E= np.append(self.E, step_to_append.E)
         self.i=np.append(self.i, step_to_append.i)
         dt = (step_to_append.setup_data.dateTime-self.setup_data.dateTime) / np.timedelta64(1,'us') /1.0e6
-        print("dt", dt , len(step_to_append.Time+ dt))
+        # print("append_dt", dt , len(step_to_append.Time+ dt))
         step_to_append.Time[0]=np.nan
         self.Time=np.append(self.Time, (step_to_append.Time+ dt ))
         
@@ -382,8 +406,10 @@ class Step_Data(EC_Setup):
         
         last_Active_RE = self.setup_data.getACTIVE_RE()
         end_norm_factor = EC_Setup.set_active_RE(self, shift_to)
-    
-        self.E_label = "E vs "+ self.setup_data.getACTIVE_RE()      
+        baseLabel = "E"
+        if self.IR_COMPENSATED:
+            baseLabel = "E - iR"
+        self.E_label = f"{baseLabel} vs "+ self.setup_data.getACTIVE_RE()      
         if end_norm_factor is not None:
             if  self.E_shifted_by == end_norm_factor.value :  
                 pass #potential is already shifted.
@@ -394,7 +420,7 @@ class Step_Data(EC_Setup):
                     self.E_shifted_by = end_norm_factor.value
                     self.E = self.E - self.E_shifted_by
                 else:
-                    self.E_label = "fdsdE vs "+ self.RE
+                    self.E_label = f"{baseLabel} vs "+ self.RE
                     self.E_unit = self.E_unit = "V" 
                     self.E = self.E + self.E_shifted_by
                     self.E_shifted_by = None   
@@ -459,7 +485,7 @@ class Step_Data(EC_Setup):
         array_Q = integrate.cumulative_simpson(step.i[idxmin:idxmax], x=step.Time[idxmin:idxmax], initial=0)
         Charge = QV(array_Q[len(array_Q)-1]-array_Q[0],step.i_unit.replace("A","C"),self.i_label.replace("i","Q")) #* QV(1,"s","t")
         
-        options = plot_options(kwargs)
+        options = plot_options(**kwargs)
         options.options["plot"] = analyse_plot
         options.x_data = step.Time[idxmin:idxmax]
         options.x_label, options.x_unit = "t", "s"
@@ -472,6 +498,24 @@ class Step_Data(EC_Setup):
         return Charge
     
     
+    def export_to_lsv(self,step_nr, *args, **kwargs):
+        """Export the data to a lsv file.
+
+        Args:
+            step (_type_): _description_
+        """ 
+        from .lsv_data import LSV_Data
+        
+        lsv = LSV_Data()
+        step = self.get_step(step_nr)
+        lsv.setup_data = copy.deepcopy(self.setup_data)
+        lsv.setup_data.name =str(self.setup_data.name)  + '#' + str(step_nr)
+        #lsv.setup_data.dateTime = self.setup_data.dateTime + np.timedelta64(int(self.Time[0]*1000000),'us')
+        #print("step", step.E[len(step.E)-1])
+        #print("step", step.E)
+        lsv.convert(step.Time, step.E,step.i,step.E[0],step.E[len(step.E)-1])
+        return lsv
+            
     def Tafel(self, lims=[-1,1], *args, **kwargs):
         x_data =np.empty(self.nr_of_steps)
         y_data =np.empty(self.nr_of_steps)